+++ src/main/java/com/dgbuild/watermyplants/services/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.dgbuild.watermyplants.services;
 
+import com.dgbuild.watermyplants.exceptions.AccessDeniedException;
 import com.dgbuild.watermyplants.exceptions.ResourceNotFoundException;
 import com.dgbuild.watermyplants.models.Plant;
 import com.dgbuild.watermyplants.models.Role;
@@ -36,9 +37,10 @@ public class UserServiceImpl implements UserService{
     @Override
     public void delete(long id) {
         userRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("User " + id + " Not Found"));
-//        if (helperFunctions.getCurrentAuditor() != "SYSTEM" || !helperFunctions.isAuthorizedToMakeChange(userRepository.findById(id).get().getUsername())){
-//            throw new OAuth2AccessDeniedException("You cannot delete other users");
-//        }
+        if(!helperFunctions.isAuthorizedToMakeChange(userRepository.findById(id).get().getUsername())){
+            throw new AccessDeniedException("You cannot delete other users");
+        }
+
         userRepository.deleteById(id);
     }
 
@@ -55,7 +57,7 @@ public class UserServiceImpl implements UserService{
 
         newUser.setEmail(user.getEmail());
 
-        newUser.setPassword(user.getPassword());
+        newUser.setPasswordNoEncrypt(user.getPassword());
 
         newUser.setPhone(user.getPhone());
 
@@ -75,10 +77,9 @@ public class UserServiceImpl implements UserService{
 
             newUser.getRoles().add(new UserRoles(newUser, addRole));
         }
+
         newUser = userRepository.save(newUser);
-//        if (helperFunctions.getCurrentAuditor() != "SYSTEM" || !helperFunctions.isAuthorizedToMakeChange(newUser.getUsername())){
-//            throw new OAuth2AccessDeniedException("You cannot change other users");
-//        }
+
         return newUser;
     }
 
@@ -91,6 +92,10 @@ public class UserServiceImpl implements UserService{
                 .orElseThrow(() -> new ResourceNotFoundException("User " + id + " Not Found"));
         newUser.setUserid(id);
 
+        if (!helperFunctions.isAuthorizedToMakeChange(userRepository.findById(id).get().getUsername())){
+            throw new AccessDeniedException("You cannot change other user's information");
+        }
+
         User currentUser = userRepository.findById(id).get();
 
         if (updateUser.getEmail() != null){
@@ -100,9 +105,9 @@ public class UserServiceImpl implements UserService{
         }
 
         if (updateUser.getPassword() != null){
-            newUser.setPassword(updateUser.getPassword());
+            newUser.setPasswordNoEncrypt(updateUser.getPassword());
         }else{
-            newUser.setPassword(currentUser.getPassword());
+            newUser.setPasswordNoEncrypt(currentUser.getPassword());
         }
 
         if (updateUser.getPhone() != null){
@@ -141,9 +146,6 @@ public class UserServiceImpl implements UserService{
             newUser.getRoles().add(new UserRoles(ur.getUser(), ur.getRole()));
         }
 
-//        if (helperFunctions.getCurrentAuditor() != "SYSTEM" || !helperFunctions.isAuthorizedToMakeChange(newUser.getUsername())) {
-//            throw new OAuth2AccessDeniedException("You cannot change other users");
-//        }
             return userRepository.save(newUser);
     }
 
@@ -167,61 +169,4 @@ public class UserServiceImpl implements UserService{
         return currentUser;
     }
 
-//    @Override
-//    public User updateCurrentUser(User newInfo) {
-//        User currentUser = userRepository.findByUsername(helperFunctions.getCurrentAuditor());
-//        User newUser = new User();
-//
-//        if (newInfo.getEmail() != null){
-//            currentUser.setEmail(newInfo.getEmail());
-//        }else{
-//            newUser.setEmail(currentUser.getEmail());
-//        }
-//
-//        if (newInfo.getPassword() != null){
-//            newUser.setPassword(newInfo.getPassword());
-//        }else{
-//            newUser.setPassword(currentUser.getPassword());
-//        }
-//
-//        if (newInfo.getPhone() != null){
-//            newUser.setPhone(newInfo.getPhone());
-//        }else{
-//            newUser.setPhone(currentUser.getPhone());
-//        }
-//
-//        if (newInfo.getUsername() != null){
-//            newUser.setUsername(newInfo.getUsername());
-//        }else{
-//            newUser.setUsername(currentUser.getUsername());
-//        }
-//
-//        for (Plant p : currentUser.getPlants()) {
-//            newUser.getPlants().add(p);
-//        }
-//
-//        for (Plant p : newInfo.getPlants()) {
-//            newUser.getPlants().add(new Plant(p.getNickname(), p.getSpecies(),
-//                    p.getFrequency(), currentUser));
-//        }
-//
-//        newUser.getRoles().clear();
-//
-//        Set<UserRoles> URSet = new HashSet<>();
-//        newInfo.getRoles().iterator().forEachRemaining(URSet::add);
-//        newInfo.getRoles().clear();
-//
-//        newUser.getRoles().clear();
-//
-//        for (UserRoles ur: currentUser.getRoles()){
-//            newUser.getRoles().add(ur);
-//        }
-//        for (UserRoles ur: newInfo.getRoles()){
-//            newUser.getRoles().add(new UserRoles(ur.getUser(), ur.getRole()));
-//        }
-//
-//        if (helperFunctions.isAuthorizedToMakeChange())
-//
-//        return userRepository.save(newUser);
-//    }
 }
